Replaced cockpit clock and changed AAR light logic
Replaced the existing cockpit clock with one modeling the Dodane
Type 11 with a roatating bezel and a cronograph (that I was sorely
missing when doing dead reckoning navigation).

The textures and lighting can be improved and there is some smaller
issues with the 3D model.  Currently it will only show time based on
/instrumentation/clock/indicated-time-sec and can not be set to
another time.

The air-to-air refueling status light logic has been changed so that
both lights will be out when tanks are full.  This mimics the
behaviour the signal lights on most drogue and probe refueling pods.

Signed-off-by: Johan Gustafsson <view.o.matic@gmail.com>